Fix all clang warnings.

+++ src/graphics.h
@@ -9,7 +9,7 @@ struct vertex {
 };
 
 struct vertices {
-    unsigned int vert_count;
+    unsigned long vert_count;
     struct vertex verts[];
 };
